refactor: unify websocket service
