+++ frontend/src/app/components/logs/logs.component.ts
@@ -1,7 +1,7 @@
 import { Component, ElementRef, ViewChild, NgZone } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AppMaterialModule } from '../../app.module';
-import { WsService } from '../../services/ws.service';
+import { WsService } from '../../core/services/ws.service';
 
 interface LogRow {
     ts: number;
